refactor: add-runtime flags mirror spec schema, remove toX mappers
